Add files via upload
Solve minor issue that when overwriting a file with same name (either in config folder or queue folder) will raise error. Now the case-sensitive will be checked, file be removed and then add the new updated file to overwrite the file name (windows being case-insensitive so it will not update the file name unless removed).